config: accept value-less keys in configuration files
If stg is run with a configuration file which has a key defined without
a value, the following exception is produced:

Traceback (most recent call last):
  File "./stg", line 24, in <module>
    from stgit.main import main
  File "stgit/main.py", line 86, in <module>
    cmd_list = stgit.commands.get_commands()
  File "stgit/commands/__init__.py", line 61, in get_commands
    for mod, m in _find_commands())
  File "stgit/commands/__init__.py", line 60, in <genexpr>
    return dict((text(getattr(m, 'name', mod)), (mod, _kinds[m.kind], m.help))
  File "stgit/commands/__init__.py", line 45, in _find_commands
    m = get_command(mod)
  File "stgit/commands/__init__.py", line 29, in get_command
    return __import__(__name__ + '.' + mod, globals(), locals(), ['*'])
  File "stgit/commands/diff.py", line 48, in <module>
    ] + argparse.diff_opts_option()
  File "stgit/argparse.py", line 208, in diff_opts_option
    default = (config.get('stgit.diff-opts') or '').split(),
  File "stgit/config.py", line 67, in get
    self.load()
  File "stgit/config.py", line 63, in load
    key, value = line.split('\n', 1)
ValueError: need more than 1 value to unpack

Git considers this valid (as when parsing booleans, a key without
a value at all is considered as true), so we should not be producing an
exception.

To fix this, catch the ValueError, and use None as the value. This
should be fine, since looking up such a key produce None, which is the
expected behavior for get(), getall(), and getint().

We choose this route instead of simply leaving the key undefined, (which
would thus produce a KeyError exception upon lookup), as we will need to
distinguish this case separately in a following patch which adds
getbool() support.

Add a new test-case which highlights the issue.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>