Repair pager for Python2.6
The pager used with `stg show`, `stg patches`, and `stg diff` comes from
the pydoc module in the Python standard library.

For Python 2.6, the pager attempts to encode the unicode text as ascii and
thus any non-ascii text causes the pager to crash.

To repair, we adopt the strategy used in Python 2.7: encode the text to be
paged using the locale's preferred encoding and 'xmlcharrefreplace' error
handling.

Repairs #19.

Signed-off-by: Peter Grayson <jpgrayson@gmail.com>