Remove Repr class; implement __repr__()
Various classes in stgit.lib.git were inherting the Repr class which
defined __repr__() in terms of __str__(). These same classes implemented
__str__(), but not __repr__(). The net result for all affected classes was
that instances' repr() and str() representations were the same.

The easier/canonical path to achieving this result is for a class to just
implement __repr__(). By default, an instance's repr is used as its str
when __str__() is not implemented.

This repairs a problem with TreeData and BlobData that could come up when
debugging. Those classes inherited Repr which defined __repr__() in terms
of str(), but those classes did not implement __str__(). This led to the
problem where calling either str() or repr() on a TreeData or BlobData
instance would lead to infinite recursive calls:

    str(repr(str(repr(str(...)))))

Signed-off-by: Peter Grayson <pete@jpgrayson.net>