Add flake8 lint support
This is the first step to introduce linting to the stgit repo using flake8.

A setup.cfg file is added with flake8 config section. Currently none of
flake8 lints are selected (i.e. `select =`).

The Makefile gains a lint target that runs flake8.

The travis configuration also now runs flake8 for the Python 2.7 and 3.7
builds.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>