More version finding repair
When setup.py is run from an sdist distribution, it is expected that the
version will come from stgit/builtin_version.py which was generated when
the sdist was built. Thus setup.py is updated to look for the builtin
version along with the describe and archival versions.

Also repair stgit.version.git_archival_version() to strip the leading 'v'
from the tag to get the version-proper.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>