stg show error checking and test cases
The error checking of `stg show` command line options is improved.

It is now an error for the user to specify both --applied and --unapplied.
Previously, if both were specified then --applied would take effect.

Similarly, it is now an error for the user to specify individual patch
names when either --applied or --unapplied is specified. Previously such a
command line would have resulted in the patches being ignored and --applied
taking effect.

The usage strings are updated to reflect these changes.

A new t0005-show.sh test module is added with several basic tests.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>