test.py is more flexible with test names
When in the working tree root, one might want to run a specific test case
using test.py. It is natural to want to type, for example:

    python t/test.py t/t0000-dummy.sh

But that form was not accepted since test.py looked for tests relative to
its directory (i.e. relative to t/). Thus the following was required:

    python t/test.py t0000-dummy.sh

With this change, both forms work.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>