Handle Run failure error data with no encoding
If `Run(...).decoding(None)` has an error (return code is non-zero), an
exception will occur in Run.__run_io() because of an attempt to decode
the stderr data (bytes) with an encoding of `None` instead of a valid
encoding. This error prevents the stderr from the failed command to be
shown to the user.

The repair is to expose a MessagePrinter.err_bytes() method that will write
the encoded bytes from the failed Run command to stgit's error stream. This
new err_bytes() method complements the existing stdout_bytes() method.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>