Repair `stg patches --diff` output
When running `stg patches --diff`, the bytes output from git.diff() was
being interpolated with the diff_tmpl str. This resulted in the diff bytes
appearing in their repr for, i.e. `b'...'`, in the user visible output.

The diff_tmpl is retired in favor of producing equivalent bytes without
using interpolation.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>