Compatible use of iconv in tests
The iconv implementation found in busybox does not comprehend the `-o`
option nor the long `--from-code` or `--to-code` options. For
compatibility, we now redirect stdout to a file instead of using `-o` along
with using short `-f` and `-t` options.

Also, the 'UTF-32' encoding is not handled well by musl libc. It does,
however, handle the more specific 'UTF-32BE' and 'UTF-32LE' encodings. The
tests in t4400-encodings.sh are modified to use 'UTF-32BE' instead of
'UTF-32'.

  https://wiki.musl-libc.org/functional-differences-from-glibc.html

Signed-off-by: Peter Grayson <pete@jpgrayson.net>