Move templates to stgit package
The templates directory moves into the stgit package. This makes the search
order for the templates cleaner (templates.py). Previously, the parent of
the stgit package would be searched for packages. This worked well when
running stg from a working tree, but meant that an installed stg would end
up searching, for example, /usr/lib/python3.7/site-packages/templates for
templates.

The templates are now installed within the stgit package directory and can
thus be reliably found without going "out of bounds" of the stgit
installation.

The templates are also still installed to share/stgit/templates, so users
can still find them.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>