Improve Makefile's coverage targets
The `coverage` target was previously broken when running parallel make
(e.g. `make -j4`) because doing so would cause the coverage-report target
to run in parallel with the coverage-test target.

The new dependency structure allows `make -j4 coverage` to correctly run
all of the tests in parallel with the coverage-report target only being run
after tests complete (i.e. after the .coverage file is complete).

Another repair is to remove the build directory before running coverage.
This ensures that the shebang in the builds is regenerated to reflect the
current python environment. This is relevant if the coverage package is not
installed globally, but instead comes from a virtual environment.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>