Reproducible bash completion generation
The generated bash completion script was different from build to build due
to the arbitrary ording of sets used in the Compgen classes. These build to
build differences caused stgit's build to not be reproducible.

The repair is to modify the Compgen classes to use regular lists instead of
sets. Although the union operations are now theoretically more expensive,
in practice these lists have a small handful of items so no material effect
on runtime performance is expected.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>