Format version 3: capture branch protect in config
A branch can be 'protected' using `stg branch --protect`. This prevents
certain modifications to the branch such as from `stg branch --delete` or
`stg pull`.

In format version 2, the protect state of a branch was determined by the
existence (or not) of a file: .git/patches/<branch>/protect.

Using .git/patches for stgit state is deprecated. So as a small step
towards migrating to the "lib" infrastructure, the branch protect state is
now captured in the `branch.<branch>.stgit.protect` config key. This is the
same config group where `stackformatversion` is captured.

There was a choice as to whether the branch protection state moved to the
config or whether it moved into the metadata file (which is committed to
the `<branch>.stgit` branch). The rationale for putting the branch
protection state in the config is that it is a local, per-repo decision
whether to protect a stgit branch. Ultimately, when the the old
.git/patches state is fully replaced by the .stgit metadata branch, it will
be possible to push and pull stgit metadata to a remote repo. When pulling
a stgit stack from a remote, the decision as to whether the branch should
be protected is unrelated to whether that branch on the remote was
protected. I.e. the local branch's protected state does not need to be
derived from the remote branch.