Migrate `stg pull` and `stg rebase` to lib infra
The pull and rebase commands now use the new lib infrastructure. Patches
are now popped and reapplied using StackTransaction for both pull and
rebase.

N.B. previously, post_rebase() would write an `orig-base` file containing
the current head id to `.git/patches/<branch>`. This file is no longer
written nor is this informationi captured elsewhere. The orig-base file was
not referenced for any other operations--the only other reference to it is
to delete it when deleting a branch.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>