Repair picking commits without names
When picking an arbitrary commit, if the user did not explicitly provide a
patch name (i.e. with --name/-n), stgit would crash due to None being
propogated as the patch name.

The make_patch_name() utility function is now used to generate a patch name
when the user does not provide one. This works whether or not the picked
commit has an empty or non-empty message.

New test cases are added.

Repairs #39.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>