Include documentation build in coverage
Some of stgit's code is used exclusively during the documentation build
process; specifically for generating asciidoc .txt files.

The stg-build tool is modified so that, like stg, coverage data will be
captured when the COVERAGE_PROCESS_START env variable is set.

A new build-txt target is added to Documentation/Makefile so that stg-build
can be invoked, but without needing to invoke asciidoc itself as the html
and man targets do. This avoids needing to install asciidoc in the CI
environment.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>