Do not use dunder prefix for attribute names
Instance attribute names using a dunder ('__') prefix can make debugging
more difficult due to how Python mangles such names at runtime. And as used
in stgit, these dunder prefixed attributes do not fit the nominal use case
for dunder prefixes to help with name clashes between super and sub
classes:

https://docs.python.org/3/reference/lexical_analysis.html#reserved-classes-of-identifiers

Some instance attributes with dunder prefixes are renamed to use a single
'_'. This still signals to readers that the attribute is private, but
allows the variable to accessed with less trouble in a debug context.

Some other instance attributes with dunder prefixes that serve to back
properties are replaced along with the property with regular (non-property)
attributes.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>