Stat instead of full diff in per-patch metadata
The per-patch metadata stored in the stack log branch (e.g. patches/p0 in
the master.stgit branch) is only consumed by humans. As such, inserting the
full diff of the patch into the metadata file is not required for stgit to
function.

In #44, attempting to refresh a very large patch affecting many files
caused an apparent hang. Serializing a very large diff can consume an
overwhelming amount of CPU an memory resources. It also puts a large
magnitude of data in the stack log branch (and thus in the repo), which can
be considered wasteful.

Instead of the full diff, only the diff stat is inlined in the per-patch
metadata. With this change, `stg refresh` has been measured to be an order
of magnitude faster for very large patches.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>