Update `stg rename` to lib infra
Stack, Patches, and Patchorder are updated with new methods to help with
patch renaming.

StackTransaction is not used to conduct the rename. The problem with using
StackTransaction for rename is that bad patch names are not discovered
until StackTransaction.run() which can leave the stack in an inconsistent
state. This forces all the error checking into commands/rename.py.

So instead Stack.rename_patch() takes responsibility for either conducting
the rename atomically or raising an appropriate error.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>