Repair branch create setting parent remote
The when creating a branch from a parent committish, the parent's remote is
meant to be copied to the new branch via the "branch.<name>.remote" config
key.

The git.identify_remote() function was failing to properly identify the
parent's remote and thus the new branch's remote was not being set. This
was largely due to the sections_matching() function using incorrect quoting
such that passed-in regular expressions would never match. But also
identify_remote() was going through a lot of complicated rigamarole,
including looking in legacy .git/remotes directory, to try and find the
parent's remote.

We abandon that old complexity and simply look for the parent's remote with
the "branch.<name>.remote" config item.

A variety of tests are also added and updated.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>