Add --keep-cr option to stg import
This option maps to the --keep-cr option of `git mailsplit` and may be used
to retain CRLF line endings when importing patches from emails.

Since RFC-compliant emails are supposed to use CRLF line endings, the
default behavior of git and stgit (and Python's email module) is to strip
the CR characters such that the mail content (message and patch) have LF
endings.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>