Remove extra options from auto_edit_patch()
The message passed into auto_edit_patch() never contains a diff, so the
contains_diff option was unnecessary.

Also, there is no command line option to modify the committer, so that
argument to auto_edit_patch() was also a no-op.

And since auto_edit_patch() does not deal with a diff that might fail to
apply, it follows that interactive_patch does not ever have a
replacement_diff.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>