Binary parsing of git commit data
This is a step toward more complete/correct handling of git commit message
data.

Instead of immediately decoding all the commit data from cat-file as utf-8,
the header data is now parsed as bytes. Git guarantees that the header
information is ascii-encoded, but the commit message following the headers
may use any arbitrary encoding.

With this change, the message data is still decoded as utf-8, but only
after the message is split from the headers.

Also now only construct a CommitData object *after* parsing all headers.
This will trigger a failure if any of the headers are missing. 

Signed-off-by: Peter Grayson <pete@jpgrayson.net>