Rework CommitData and Person initialization
The CommitData and Person classes' initialization is updated to remove the
complexity of make_defaults() and NoValue. Instead of indirecting __init__
arguments through a function created by make_defaults(), instance
attribute assignment is now done in a more straightforward manner in
__init__.

The new _replace() method is used to implement the various setter methods.
It constructs new CommitData and Person copies with specific fields
replaced.

N.B. CommitData now require non-None tree, parents, and message to
instantiate an instance. Only a couple of minor changes were needed to
ensure this new invariant of CommitData.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>