Deprecate stg publish
The `stg publish` command is believed to be almost/entirely unused.

The publish command is being deprecated with the intention of removing it
entirely in a future StGit version.

Removing the publish command will make StGit a bit easier to maintain. It
will also help new users by having one less complicated sub-command to
learn about.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>