Remove test.py test runner
The test.py runner has similar capabilities to prove, but not many
advantages. So to reduce the amount of code to maintain, we drop the custom
test.py script and instead recommend using 'make prove'.

N.B. setting STG_PROVE_OPTS=-j4 is required for prove to run tests in
parallel.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>