Do not fiddle with sys.path
In the stg script, there is support for dynamically supporting arbitrary
installation paths by detecting whether the installation prefix differs
from the nominal sys.prefix and, if different, inserting additional library
search paths into sys.path.

This capability is removed from stg.

The stance is that it is an install-time concern to ensure StGit is
installed to a path that will be available in sys.path at runtime; and that
it is thus out of scope for the stg script to try to fix-up sys.path at
runtime to accommodate any arbitrary installation path.

It is recommended to install StGit to one of the system paths (/usr,
/usr/local), the systemd file-hierarchy specified $HOME/.local, or into a
Python virtualenv. For StGit development purposes, no installation is
required--the stg script can be run in-place.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>