Remove conflict_old from test-lib-functions.sh
The conflict_old test wrapper was a stop-gap to accommodate stg commands
using the old git infrastructure that returned 2 instead of 3 when a merge
conflict is encountered.

There were only a couple of test sites using `conflict_old`. These sites
are modified to now use `conflict` instead.

This is accompanied by a changes to sync, pull, and main to raise and
handle MergeConflictException in a consistent manner.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>