Add configuration for black code formatter
The black code formatter is gaining traction. It enforces a reasonable
style with minimal configuration. Using black with StGit will allow code
styling to be automated and thus be one less thing to deal with when
maintaining the code.

https://github.com/psf/black

This adds the pyproject.toml file which contains black's configuration. The
only non-standard configuration option is use of skip-string-normalization.
By default, black normalizes all string literals to use double-quotes.
Enabling skip-string-normalization allows StGit to use a heterogenous mix
of single and double quotes, which reduces the noise in the reformat diff
and enables continued use of StGit's preferred single-quoting style.

Makefile, flake8, isort, and GitHub Actions are updated to incorporate
black.

A separate commit will actually reformat Python code with black.