Python 3.5 minimum version, explicitly drop 2.x
All Python versions < 3.5 are end-of-lifed by the Python project [0]. The
oldest non-EOL version is 3.5. To ease maintenance, StGit will only support
non-EOL versions of Python.

This change explicitly drops support for 2.x by virtue of removing
compatibility code. And Python versions < 3.5 are unsupported in that those
versions are no longer tested with StGit.

The `# -*- coding: utf-8 -*-` encoding declarations are removed. UTF-8 is
the default encoding for Python 3 [1], so these declarations become
superfluous.

[0] https://devguide.python.org/devcycle/#end-of-life-branches
[1] https://www.python.org/dev/peps/pep-3120/

Signed-off-by: Peter Grayson <pete@jpgrayson.net>