Use stg branch in bash and fish completions
Instead of the bash completion script inspecting .git/refs/patches to find
patches, it now runs `stg series`. The .git/refs/patches directory is an
implementation detail and will likely change soon, so using the stg
porcelain keeps the script at the right level of abstraction.

Similarly, both the bash and fish completions are also updated to use `stg
branch` to generate the list of StGit branches instead of searching
.git/refs/heads for `.stgit` log branches. The log branches may be renamed
in the future.

Finally, a few additional changes are made to the bash completions to
cleanup some shellcheck warnings.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>