Apply isort lint to all python code
The python files in the perf/ directory were not being checked by isort
when running `make lint`.

The Makefile is updated to run isort against `.` to be more inclusive. This
is also consistent with how flake8 is run for lint.

The skip_glob setting is added to setup.cfg to avoid isort descending into
any dot directories, any of which are not part of the stgit worktree.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>