Repair importing mail with colon-space in subject
The subject needs to be split exactly once, on the first colon, otherwise
any additional colons in the subject will cause the construction of the
mailinfo dict to fail.

The mbox import tests are updated to include a multi-colon subject that
will fail without this repair.

Repairs #53.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>
Thanks: Dave Kleikamp for the report and repair suggestion!