Use coverage contexts
Coverage contexts are supported as of coverage.py version 5.0.

When running code coverage (`make coverage`), the context of the coverage
data is now recorded in the coverage data files. For `stg` invocations, the
subcommand (e.g. 'refresh', 'new', etc.) is captured as the relevant
context.

For each covered line, the coverage html report shows the coverage contexts
that executed the line of code. This makes it possible to know which StGit
subcommands are using which parts of the stgit code base.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>