Limit mail diffstat to width of 72
This is similar to the policy of `git format-patch` which also limits the
diffstat in the cover letter to 72 columns.

Any user-supplied diff options (i.e. from --diff-opts) are no longer
applied to the cover letter's diffstat. Those diff opts are now only
applied to the regular diffs in the per-patch messages.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>