Repair mail SSL check
The previous `not hasattr(socket, 'ssl')` test was an invalid way to test
whether SSL support is built into the Python interpreter.

https://bugzilla.redhat.com/show_bug.cgi?id=1553834
https://github.com/zopefoundation/zope.sendmail/issues/9

New approach is to unconditionally call starttls() and catch and report the
potential errors. A RuntimeError will be raised if Python lacks SSL
support. This also improves the error reporting if the server does not
handle STARTTLS, wherein an SMTPException will be raised.

Repairs #57.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>