Repair stg mail with both -a and -e options
Using both -a and -e (--all and --edit-cover) with `stg mail` triggered an
odd corner case where the args passed to __send_message() would be the list
of patches which caused a string interpolation to fail.

The repair is to make __send_message() somewhat less clever such that it
either handles msg_type == 'cover' or msg_type == 'patch', but not parts of
both simultaneously.

A test is added for this issue.

Repairs #58.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>