Quote whole command in man page synopsis
When generating the manpage asciidoc, in the SYNOPSIS section, the full stg
subcommand is now quoted instead of just quoting stg. E.g.:

    'stg diff'

instead of:

    'stg' diff

The net of this is that the full command is emphasized in the rendered
output. I.e. either italicized (html) or bold (man).

This change is consistent with Git's manpages' asciidoc.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>