Repair diffstat when outside work tree root
When invoking `stg mail` or `stg edit --diff` from a subdirectory of the
work tree root, the diffstat would exclude any files not relative to the
subdirectory.

This was due to calling `git apply --stat` from the current directory
instead of the wort tree root. The solution is to move diffstat() to be a
method of IndexAndWorktree such that `git apply --stat` is always run from
the worktree root.

Repairs #62.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>