Windows: Pass environment variables to Popen as dict of strings
In Python 3, the env argument to Popen has to be a dict of strings, see
https://bugzilla.mozilla.org/show_bug.cgi?id=1585258#c4
This is important when running under Windows, where the unicode strings
are processed directly by the subprocess C code.

However, under posix systems the environment variables are serialized to
bytes using the file system encoding, which may not support unicode
characters.
This means that passing bytes as environment variables and skipping the
encoding is mandatory on posix in order to support unicode environment
variables.

Therefore skip encoding when on Windows.