Better stack rename abstraction
The new Refs.rename() method encapsulates the execution of `git
update-ref` such that Stack.rename() no longer executes update-ref
directly.

Stack.rename() is further updated to incorporate the rename of the
metadata branch (i.e. <branch>.stgit). This was previously performed in
commands/branch.py, but since the metadata branch is part-and-parcel to
a Stack, it makes sense that a Stack rename must also rename its
associated metadata ref.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>