Repair `stg log` with patches from subdir
If `stg log` was run from a worktree subdir (i.e. not from the root), and
if patches were provided on the command line, nothing would be printed.
This is due to the patch names being mapped to pathlimits to the underlying
`git log` (or `gitk`) commands.

This is resolved by using Repository.run(), which takes care to operate
from the root, instead of plain Run, which runs from the current directory.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>