Repair potential hang in DiffTreeProcess
When calling DiffTreeProcess.diff_trees(), it was possible for that
method to hang forever. This would happen when the 'EOF\n' terminator
would happen to be split across os.read() calls, which could rarely
happen when the diff was larger than io.DEFAULT_BUFFER_SIZE.

This issue was observed when using `stg pull -m`, `stg goto -m`, and
`stg push -m`.

Signed-off-by: Pete Grayson <pete@jpgrayson.net>