Absolute Worktree.directory
Instead of Worktree.directory being the relative cdup path to the root of
the worktree, it is now the absolute path to the root of the worktree.

This allows Worktree to just have a regular env() method obviates the
env_in_cwd() method; i.e. because GIT_WORK_TREE is now always the absolute
path.

IndexAndWorktree is further simplified by removing the cwd property. The
only user of iw.cwd, get_hook(), now references Worktree.directory instead.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>