Repair crash when attempting to export empty patch
When using git apply to generate a diffstat the diff must not be empty
otherwise git exits with a failure that cascades to stg export also
failing.

Also repair importing an empty patch. There was a regression when
splitting the diff from the description which would cause stg import to
fail. The message/diff separator is now removed from the diff string.

Fixes #112

Signed-off-by: Peter Grayson <pete@jpgrayson.net>