Ensure a newline after each commit in 'stg squash'.
I observed an issue locally where if a commit message does not include a
trailing newline the suggested commit message is more "scrunched" than
we'd like it to be, with no newline between the commit message and the
instructions.

This scenario is easy to get into if you perform a squash followed by
another squash onto the same patch.

The fix here is to strip all newlines from the end of each commit
message and then add two newlines to the end (one to get to the next
line, another to ensure the blank line between messages).