Reorder squash message to ease readibility.
After using this in practice, I think that having the patch number
followed by the patch name is more readable. With this change it's
easier to scan the commit message to determine where each change
ends and the next one begins.

Before:

After:

I recognize that this PR is more personal preference; it's not the end
of the world if we leave it as-is.