Add --autostash to 'stg rebase'
This mirrors a feature of 'git rebase' where, instead of erroring on a dirty
worktree, `--autostash` will stash the changes and apply them after the rebase
has completed.

I'm not sure if it's more 'stgit-ish' to create a patch or a stash. I kept the
git behavior (a stash) but I could see arguments either way.

[![asciicast](https://asciinema.org/a/423971.svg)](https://asciinema.org/a/423971)