new: commit.verbose is a boolean
In #122, `new` got a `--verbose` flag that defaults to true if the
`commit.verbose` configuration in git is set. To determine whether the
configuration was set, the new feature used `config.getint`.

However, per the git-config documentation, `commit.verbose` can be a
boolean:

    commit.verbose
       A boolean or int to specify the level of verbose with git commit. See git-commit(1).

In fact, if `commit.verbose` is set to `true` right now, this will fail with,

    Value for "commit.verbose" is not an integer: "true"

Given that `config.getbool` will handle both, integers and booleans, use
that to determine if the configuration is set instead. Since
`config.getbool` will return `None` if unset, default to False rather than
None.