Fix off-by-one-error for 'stg rebase -i delete'
Previously, we were using itertools.count(). That's great! But we also were
performing an `index -= 1` in the loop which was getting ignored.

The new algorithm explicitly deletes the patch from the instruction set and
correctly counts index on deletes.

Added a test case which proves the off-by-one error is fixed and that deletes
work correctly.