completion: Add --no-description for stg series
If the `stg.series.description` config is set to `true`, then the output of
`stg series` always includes the first line of the commit messages.

For example,

    $ stg series
    + foo # Add foo
    + bar # Rename bar
    > baz # Remove baz

When the completion scripts call `stg series`, they use the full string,
including the description:

    $ stg refresh -p <tab>
    foo\ \#\ Add\ foo

To solve this, pass the `--no-description` flag to `stg series` in all
completion scripts when retrieving a list of patches.