Workaround child process reaping race on Windows
On Windows, there is apparently a race between StGit attempting to kill its
child CatFileProcess at exit and those processes exiting themselves when
stdin is closed, such that the CatFileProcess _shutdown() methods will
sometimes (always?) get an OSError when attempting to kill these child
processes.

Since the end-goal is for these processes to be killed, the workaround is
to simply ignore the OSError exception if os.kill() fails.

Closes #78