Do not close stdin before killing child bg process
Addresses #78

Closing stdin of the child git process to terminate itself. On Windows, the
subsequent attempt to kill the already-terminating child process results in
an exception (PermissionError or WindowsError or OSError). On Linux, the
terminated child process remains in a defunct/zombie state so that the
subsequent kill is idempotent and the child is finalized by wait().

Signed-off-by: Peter Grayson <pete@jpgrayson.net>