Add push --noapply
The new --noapply option to push allows the order of unapplied patches to
be modified without actually applying those patches.

It may be helpful to some workflows to allow patches to be reordered
without either (a) having to immediately resolve merge conflicts; or (b)
having to modify the contents of the working tree.

Addresses #144.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>