Find viable bash.exe for running hooks on Windows
When running hooks on Windows, bash.exe must be supplied on the command
line to run the hook script. This change tries to find a viable bash.exe
from the user's PATH, while avoiding the bash.exe stub in the OS's System32
directory.

This change is adapted from Michael Wild's PR #103.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>