Use different dynamic version system
This is a total overhaul of the previous dynamic version system. A detailed
theory of operation can be found in _version.py. Key features include:

- Obtain stgit version dynamically from `git describe` when operating from
  within a stgit git working tree.
- Obtain stgit version statically from stgit/_static_version.py when
  operating from a stgit installation or stgit distribution such as an
  sdist tarball or git archive tarball.
- Minimum overhead--never attempt to run git when in the context of a stgit
  distribution or installation.
- Play well with setuptools and pip.
- No third-party dependency**.

** N.B. This code is adapted from the CC0 1.0 (public domain) licensed
miniver project. A challenge with code of this sort is that it must reside
in-tree and be woven between build-time (setup.py) and runtime (stgit
package). The _version.py file is replaced in sdist code such that it only
contains a static version string and not the code needed to dynamically
determine the version. In this manner, the public domain code is not
distributed in StGit's source distributions.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>