Ensure coveralls uses coverage[toml]
A problem emerged where `make dev-env` installs the latest version of
the coverage package (6.x) with [toml] specified for its extras, but
then installing coveralls, which currently depends on coverage<6.0,
downgrades coverage, but does not install the [toml] extras in the
downgraded version. This seems like a bug in pip since it should
ostensibly preserve extras when downgrading.

We work around the issue by re-specifying "coverage[toml]" when
installing coveralls so that if coverage is downgraded, the [toml]
extras are still installed.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>