Update tests for updated git pull semantics
As of approximately Git 2.27, using `git pull` without any of the
pull.rebase or pull.ff config options or the --rebase=false command line
option will cause the pull to abort with a warning message. The test
suite is updated to set pull.rebase=false, which was the default
behavior, in various contexts that require it.

Also, pull semantics have also changed such that the pull strategy
used in t2110-pull-stack.sh is invalidated. Those tests are modified to
use a more robust fetch and reset strategy.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>