Proper implementation for `stg id`
The user-provided StGit revision is parsed

Add -b/--branch option for id. This is a new feature compared to the
Python version.

Another difference vs Python version: the provided revspec is not always
peeled to a commit object. This specifically affects the output of
annotated tags where, e.g. `stg id v1.3` shows the id of the commit the
tag points to where as `git rev-parse v1.3` shows the id of the tag
itself.