Implement patch ranges for `stg series`
The user may now specify the patch range(s) to be shown via patch range
arguments.

This implementation differs from the Python version in that the patch
ranges specified must be (a) in-order and (b) contiguous. The Python
version did not impose either of those constraints, which allowed for
some hilariously wrong/misleading outputs.

New tests are implemented.