Don't use wrapped Repository or Signature
More trouble than it's worth. The only Repository function that
currently needs different behavior is commit(), so that wrapped function
(no longer associated with a wrap::Repository struct) is what remains as
commit_ex().

Also fall-back to plain git2::Signature's everywhere. Although StGit can
enforce its own commit metadata to be utf-8, it still needs to be able
to gracefully handle commits with non-utf-8 or otherwise broken
encodings when doing repair, import, uncommit, etc. So this means that
the complexities of non-utf-8 signatures cannot be hidden away and must
be comprehended everywhere.